キャラクター関係で必要そうなクラスの作成
