+++ MMOProduction/Assets/Script/Character/CharacterBase.cs
@@ -12,18 +12,10 @@ public abstract class CharacterBase : MonoBehaviour
     protected string characterName = "";
     public string Name { get { return characterName; } set { characterName = value; } }
 
-    protected float nowFlame = 0;
-
     protected int id = 0;   // キャラクターごとのID
-
-    protected Vector3 lastPos = new Vector3();
-    protected Vector3 nextPos = new Vector3();
-    protected Quaternion lastDir = new Quaternion();
-    protected Quaternion nextDir = new Quaternion();
 
     protected Animator animator_;
 
-    public const float UPDATE_SPEED = 1.0f / 3.0f;
 
     /// <summary>
     /// キャラクターの初期化設定
@@ -32,46 +24,15 @@ public abstract class CharacterBase : MonoBehaviour
     /// <param name="_y"></param>
     /// <param name="_z"></param>
     /// <param name="_dir"></param>
-    public void Init(float _x, float _y, float _z, float _dir)
-    {
-        lastPos = transform.position = new Vector3(_x, _y, _z);
-        nextPos = new Vector3(_x, _y, _z);
-        lastDir = transform.rotation = Quaternion.Euler(new Vector3(0, _dir, 0));
-        nextDir = Quaternion.Euler(new Vector3(0, _dir, 0));
-        animator_ = GetComponent<Animator>();
-    }
+    //public void Init(float _x, float _y, float _z, float _dir)
+    //{
+    //    lastPos = transform.position = new Vector3(_x, _y, _z);
+    //    nextPos = new Vector3(_x, _y, _z);
+    //    lastDir = transform.rotation = Quaternion.Euler(new Vector3(0, _dir, 0));
+    //    nextDir = Quaternion.Euler(new Vector3(0, _dir, 0));
+    //    animator_ = GetComponent<Animator>();
+    //}
 
-
-    public void UpdatePostionData(float _x, float _y, float _z, float _dir)
-    {
-        // 向きを決める
-        lastDir = transform.rotation;
-        nextDir = Quaternion.Euler(0, _dir, 0);
-
-        // 位置を決める
-        lastPos = transform.position;
-        nextPos = new Vector3(_x, _y, _z);
-
-        // カウントを初期化
-        nowFlame = 0;
-    }
-
-    public void UpdateStatusData(int _hp,  int _mp, int _status)
-    {
-        hp = _hp;
-        mp = _mp;
-        status = _status;
-    }
-
-    /// <summary>
-    /// キャラクターの移動を補間する
-    /// </summary>
-    protected void LerpMove()
-    {
-        nowFlame += UPDATE_SPEED;
-        transform.rotation = Quaternion.Lerp(lastDir, nextDir, nowFlame);
-        transform.position = Vector3.Lerp(lastPos, nextPos, nowFlame);
-    }
 
     /// <summary>
     /// キャラクターごとのID
