+++ MMOProduction/Assets/Script/Character/Enemy/EnemysController.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 98b33bdcab465c94faedc3353d3e89b3
+guid: 052b21850b76c78469242a7d85695062
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
