+++ MMOProduction/Assets/Script/Character/OtherPlayer/OtherPlayersController.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: cca5fc79b88700940956820b4656ad94
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
