+++ MMOProduction/Assets/Script/Character/Resident.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 6ab3c2d5dc15f0d44980c0a1c5e62e2e
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
