+++ MMOProduction/Assets/Script/Character/Resident/Resident.cs
@@ -1,0 +1,23 @@
+﻿using System;
+using System.Collections.Generic;
+using UnityEngine;
+using System.Linq;
+using System.Threading.Tasks;
+
+/// <summary>
+/// NPC用クラス
+/// </summary>
+class Resident : NonPlayer
+{
+    // Start is called before the first frame update
+    void Start()
+    {
+    }
+
+
+    // Update is called once per frame
+    void Update()
+    {
+        
+    }
+}
