+++ MMOProduction/Assets/Script/Character/Resident/Resident.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 946a1c6bcc2e41246be42838cc5e337b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
