+++ MMOProduction/Assets/Script/Character/Resident/ResidentsController.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 7f8a80c7375ea6c48814c3628f07c2af
+guid: 3f74addc3b98e9c4a9a18b34dd974ee2
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
