+++ MMOProduction/Assets/Script/Stage/StageObject.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: c97a87026f2dbbf44a6d2ee94fc9bc4b
+guid: a04e3f014c5b5ce48a23bb99a981cb37
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
