+++ MMOProduction/Assets/Script/Targetable.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 037af1bbfadc4474cb2e3aa23162466c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
